+++ liquid-outreach.php
@@ -497,6 +497,7 @@
             $this->lo_shortcodes = new LO_Shortcodes($this);
             $this->lo_ccb_events_info_setings = LO_Ccb_Events_Info_Setings::get_instance();
             $this->lo_ccb_events_page_settings = LO_Ccb_Events_Page_Settings::get_instance();
+            $this->lo_ccb_events_partner_cat_map_settings = LO_Ccb_Events_Partner_Cat_Map_Settings::get_instance();
             
             if (is_admin()) {
                 $this->lo_ccb_api_event_profiles = new Lo_Ccb_api_event_profiles($this);